feature/dev (#21)
## 작업 요소
- User 중분류 패키지 일부 수정하였습니다.
- ChekcIn 중분류 패키지 일부 수정하였습니다.
- Reservation 중분류 패키지  일부 수정하였습니다.
- Patient 중분류 패키지  일부 수정하였습니다.

## 작업 내용
- 중분류 패키지 엔터티에 연관관계를 추가하였습니다.
 - 환자 정보를 확인하여 등록된 환자만 접수, 예약이 가능하도록 추가했습니다.
 - 연관 관계를 고려하여 코드 유기적으로 수정했습니다.
 -  테스트를 완료하였고, api 명세서를 최신화하였습니다.

## 연관 이슈
- open 

<img width="475" alt="DB" src="https://github.com/user-attachments/assets/72577e25-c1c4-43d1-8817-7d8e169d3a0a" />

+++ src/main/java/com/cloud/emr/Affair/CheckIn/dto/CheckInListResponse.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class CheckInListResponse {
 
     @NotNull(message = "접수 ID는 필수 값입니다.")
-    @Column(nullable = false, columnDefinition = "INT")
+    @Column(nullable = false, columnDefinition = "BIGINT")
     private Long checkInId;    // 접수 ID
 
     @NotBlank(message = "환자번호는 필수 값입니다.")
